+++ graspdb/include/graspdb/GraspDemonstration.h
@@ -14,16 +14,17 @@
 #ifndef RAIL_PICK_AND_PLACE_GRASPDB_GRASP_DEMONSTRATION_H_
 #define RAIL_PICK_AND_PLACE_GRASPDB_GRASP_DEMONSTRATION_H_
 
+// graspdb
+#include "Entity.h"
+#include "Pose.h"
+
 // ROS
 #include <rail_pick_and_place_msgs/GraspDemonstration.h>
+#include <sensor_msgs/Image.h>
 #include <sensor_msgs/PointCloud2.h>
 
 // C++ Standard Library
 #include <string>
-
-// graspdb
-#include "Entity.h"
-#include "Pose.h"
 
 namespace rail
 {
@@ -55,12 +56,13 @@ public:
    * \param grasp_pose The pose of the grasp (defaults to an empty Pose).
    * \param eef_frame_id The frame identifier for the end effector used (defaults to the empty string).
    * \param point_cloud The ROS sensor_msgs/PointCloud2 message of the segmented object (defaults to empty values).
+   * \param image The ROS sensor_msgs/Image message of the segmented object (defaults to empty values).
    * \param created The created timestamp (defaults to 0).
    */
   GraspDemonstration(const uint32_t id = Entity::UNSET_ID, const std::string &object_name = "",
-      const Pose &grasp_pose = Pose(), const std::string &eef_frame_id = "",
-      const sensor_msgs::PointCloud2 &point_cloud = sensor_msgs::PointCloud2(),
-      const time_t created = Entity::UNSET_TIME);
+                     const Pose &grasp_pose = Pose(), const std::string &eef_frame_id = "",
+                     const sensor_msgs::PointCloud2 &point_cloud = sensor_msgs::PointCloud2(),
+                     const sensor_msgs::Image &image = sensor_msgs::Image(), const time_t created = Entity::UNSET_TIME);
 
   /*!
    * \brief Create a new GraspDemonstration.
@@ -70,11 +72,12 @@ public:
    *
    * \param object_name The name of the object grasped.
    * \param grasp_pose The pose of the grasp.
-   * \param eef_frame_id The frame identifier for the end effector used (defaults to the empty string).
-   * \param point_cloud The ROS sensor_msgs/PointCloud2 message of the segmented object (defaults to empty values).
+   * \param eef_frame_id The frame identifier for the end effector used.
+   * \param point_cloud The ROS sensor_msgs/PointCloud2 message of the segmented object.
+   * \param image The ROS sensor_msgs/Image message of the segmented object.
    */
   GraspDemonstration(const std::string &object_name, const Pose &grasp_pose, const std::string &eef_frame_id,
-      const sensor_msgs::PointCloud2 &point_cloud);
+                     const sensor_msgs::PointCloud2 &point_cloud, const sensor_msgs::Image &image);
 
   /*!
    * \brief Create a new GraspDemonstration.
@@ -104,13 +107,22 @@ public:
   void setObjectName(const std::string &object_name);
 
   /*!
-   * \brief Grasp pose value accessor.
+   * \brief Grasp pose value accessor (immutable).
    *
    * Get the grasp pose value of this GraspDemonstration.
    *
    * \return The grasp pose value.
    */
   const Pose &getGraspPose() const;
+
+  /*!
+   * \brief Grasp pose value accessor.
+   *
+   * Get the grasp pose value of this GraspDemonstration.
+   *
+   * \return The grasp pose value.
+   */
+  Pose &getGraspPose();
 
   /*!
    * \brief Grasp pose value mutator.
@@ -140,7 +152,7 @@ public:
   void setEefFrameID(const std::string &eef_frame_id);
 
   /*!
-   * \brief Point cloud accessor.
+   * \brief Point cloud accessor (immutable).
    *
    * Get the point cloud message.
    *
@@ -149,13 +161,49 @@ public:
   const sensor_msgs::PointCloud2 &getPointCloud() const;
 
   /*!
-   * \brief Point cloud buffer mutator.
+   * \brief Point cloud accessor.
+   *
+   * Get the point cloud message.
+   *
+   * \return The point cloud message.
+   */
+  sensor_msgs::PointCloud2 &getPointCloud();
+
+  /*!
+   * \brief Point cloud mutator.
    *
    * Set the point cloud message to the given values based on the ROS message.
    *
    * \param point_cloud The ROS PointCloud2 message to store.
    */
   void setPointCloud(const sensor_msgs::PointCloud2 &point_cloud);
+
+  /*!
+   * \brief Image accessor (immutable).
+   *
+   * Get the image message.
+   *
+   * \return The image message.
+   */
+  const sensor_msgs::Image &getImage() const;
+
+  /*!
+   * \brief Image accessor.
+   *
+   * Get the image message.
+   *
+   * \return The image message.
+   */
+  sensor_msgs::Image &getImage();
+
+  /*!
+   * \brief Image mutator.
+   *
+   * Set the image message to the given values based on the ROS message.
+   *
+   * \param image The ROS Image message to store.
+   */
+  void setImage(const sensor_msgs::Image &image);
 
   /*!
    * Converts this GraspDemonstration object into a ROS GraspDemonstration message.
@@ -171,6 +219,8 @@ private:
   Pose grasp_pose_;
   /*! The point cloud data. */
   sensor_msgs::PointCloud2 point_cloud_;
+  /*! The RGB image data. */
+  sensor_msgs::Image image_;
 };
 
 }