+++ graspdb/src/GraspDemonstration.cpp
@@ -11,28 +11,32 @@
  * \date March 11, 2015
  */
 
-#include <graspdb/GraspDemonstration.h>
+// graspdb
+#include "graspdb/GraspDemonstration.h"
 
 using namespace std;
 using namespace rail::pick_and_place::graspdb;
 
 GraspDemonstration::GraspDemonstration(const uint32_t id, const string &object_name, const Pose &grasp_pose,
-    const string &eef_frame_id, const sensor_msgs::PointCloud2 &point_cloud, const time_t created)
+                                       const string &eef_frame_id, const sensor_msgs::PointCloud2 &point_cloud,
+                                       const sensor_msgs::Image &image, const time_t created)
     : Entity(id, created),
-      object_name_(object_name), grasp_pose_(grasp_pose), eef_frame_id_(eef_frame_id), point_cloud_(point_cloud)
+      object_name_(object_name), grasp_pose_(grasp_pose), eef_frame_id_(eef_frame_id), point_cloud_(point_cloud),
+      image_(image)
 {
 }
 
 GraspDemonstration::GraspDemonstration(const string &object_name, const Pose &grasp_pose, const string &eef_frame_id,
-    const sensor_msgs::PointCloud2 &point_cloud)
-    : object_name_(object_name), grasp_pose_(grasp_pose), eef_frame_id_(eef_frame_id), point_cloud_(point_cloud)
+                                       const sensor_msgs::PointCloud2 &point_cloud, const sensor_msgs::Image &image)
+    : object_name_(object_name), grasp_pose_(grasp_pose), eef_frame_id_(eef_frame_id), point_cloud_(point_cloud),
+      image_(image)
 {
 }
 
 GraspDemonstration::GraspDemonstration(const rail_pick_and_place_msgs::GraspDemonstration &gd)
     : Entity(gd.id, gd.created.sec),
       object_name_(gd.object_name), grasp_pose_(gd.grasp_pose), eef_frame_id_(gd.eef_frame_id),
-      point_cloud_(gd.point_cloud)
+      point_cloud_(gd.point_cloud), image_(gd.image)
 {
 }
 
@@ -47,6 +51,11 @@ void GraspDemonstration::setObjectName(const string &object_name)
 }
 
 const Pose &GraspDemonstration::getGraspPose() const
+{
+  return grasp_pose_;
+}
+
+Pose &GraspDemonstration::getGraspPose()
 {
   return grasp_pose_;
 }
@@ -71,9 +80,29 @@ const sensor_msgs::PointCloud2 &GraspDemonstration::getPointCloud() const
   return point_cloud_;
 }
 
+sensor_msgs::PointCloud2 &GraspDemonstration::getPointCloud()
+{
+  return point_cloud_;
+}
+
 void GraspDemonstration::setPointCloud(const sensor_msgs::PointCloud2 &point_cloud)
 {
   point_cloud_ = point_cloud;
+}
+
+const sensor_msgs::Image &GraspDemonstration::getImage() const
+{
+  return image_;
+}
+
+sensor_msgs::Image &GraspDemonstration::getImage()
+{
+  return image_;
+}
+
+void GraspDemonstration::setImage(const sensor_msgs::Image &image)
+{
+  image_ = image;
 }
 
 rail_pick_and_place_msgs::GraspDemonstration GraspDemonstration::toROSGraspDemonstrationMessage() const
@@ -84,6 +113,7 @@ rail_pick_and_place_msgs::GraspDemonstration GraspDemonstration::toROSGraspDemon
   gd.grasp_pose = grasp_pose_.toROSPoseStampedMessage();
   gd.eef_frame_id = eef_frame_id_;
   gd.point_cloud = point_cloud_;
+  gd.image = image_;
   gd.created.nsec = 0;
   gd.created.sec = this->getCreated();
   return gd;