+++ rail_grasp_collection/src/GraspCollector.cpp
@@ -10,7 +10,10 @@
  * \date March 3, 2015
  */
 
-#include <rail_grasp_collection/GraspCollector.h>
+// RAIL Grasp Collection
+#include "rail_grasp_collection/GraspCollector.h"
+
+// ROS
 #include <tf2_sensor_msgs/tf2_sensor_msgs.h>
 
 using namespace std;
@@ -54,12 +57,12 @@ GraspCollector::GraspCollector()
   // setup a debug publisher if we need it
   if (debug_)
   {
-    debug_pub_ = private_node_.advertise<sensor_msgs::PointCloud2>("debug", 1);
+    debug_pub_ = private_node_.advertise<sensor_msgs::PointCloud2>("debug", 1, true);
   }
 
   // subscribe to the list of segmented objects
   segmented_objects_sub_ = node_.subscribe(segmented_objects_topic, 1, &GraspCollector::segmentedObjectsCallback,
-      this);
+                                           this);
 
   // setup action clients
   gripper_ac_ = new actionlib::SimpleActionClient<rail_manipulation_msgs::GripperAction>(gripper_action_server, true);
@@ -185,25 +188,25 @@ void GraspCollector::graspAndStore(const rail_pick_and_place_msgs::GraspAndStore
     boost::mutex::scoped_lock lock(mutex_);
     // check if we actually have some objects
     int closest = 0;
-    if (object_list_.objects.size() == 0)
+    if (object_list_->objects.size() == 0)
     {
       as_.setSucceeded(result, "No segmented objects found.");
       return;
-    } else if (object_list_.objects.size() > 1)
+    } else if (object_list_->objects.size() > 1)
     {
       // find the closest point
       float min = numeric_limits<float>::infinity();
       //geometry_msgs::Vector3 &v = grasp.transform.translation;
       // check each segmented object
-      for (size_t i = 0; i < object_list_.objects.size(); i++)
+      for (size_t i = 0; i < object_list_->objects.size(); i++)
       {
         geometry_msgs::TransformStamped eef_transform = tf_buffer_.lookupTransform(
-            object_list_.objects[i].point_cloud.header.frame_id, eef_frame_id_, ros::Time(0)
+            object_list_->objects[i].point_cloud.header.frame_id, eef_frame_id_, ros::Time(0)
         );
         geometry_msgs::Vector3 &v = eef_transform.transform.translation;
         //convert PointCloud2 to PointCloud to access the data easily
         sensor_msgs::PointCloud cloud;
-        sensor_msgs::convertPointCloud2ToPointCloud(object_list_.objects[i].point_cloud, cloud);
+        sensor_msgs::convertPointCloud2ToPointCloud(object_list_->objects[i].point_cloud, cloud);
         // check each point in the cloud
         for (size_t j = 0; j < cloud.points.size(); j++)
         {
@@ -220,13 +223,14 @@ void GraspCollector::graspAndStore(const rail_pick_and_place_msgs::GraspAndStore
       }
     }
     // check if we need to transform the point cloud
-    rail_manipulation_msgs::SegmentedObject &object = object_list_.objects[closest];
+    rail_manipulation_msgs::SegmentedObject &object = object_list_->objects[closest];
     if (object.point_cloud.header.frame_id != robot_fixed_frame_id_)
     {
       try
       {
         sensor_msgs::PointCloud2 transformed_cloud = tf_buffer_.transform(object.point_cloud, robot_fixed_frame_id_,
-            ros::Time(0), object.point_cloud.header.frame_id);
+                                                                          ros::Time(0),
+                                                                          object.point_cloud.header.frame_id);
         object.point_cloud = transformed_cloud;
         object.point_cloud.header.frame_id = robot_fixed_frame_id_;
       } catch (tf2::TransformException &ex)
@@ -245,7 +249,8 @@ void GraspCollector::graspAndStore(const rail_pick_and_place_msgs::GraspAndStore
     // store the data
     feedback.message = "Storing grasp data...";
     as_.publishFeedback(feedback);
-    graspdb::GraspDemonstration gd(goal->object_name, graspdb::Pose(grasp), eef_frame_id_, object.point_cloud);
+    graspdb::GraspDemonstration gd(goal->object_name, graspdb::Pose(grasp), eef_frame_id_, object.point_cloud,
+                                   object.image);
     if (graspdb_->addGraspDemonstration(gd))
     {
       // store the ID
@@ -262,7 +267,7 @@ void GraspCollector::graspAndStore(const rail_pick_and_place_msgs::GraspAndStore
   as_.setSucceeded(result, "Success!");
 }
 
-void GraspCollector::segmentedObjectsCallback(const rail_manipulation_msgs::SegmentedObjectList &object_list)
+void GraspCollector::segmentedObjectsCallback(const rail_manipulation_msgs::SegmentedObjectList::Ptr &object_list)
 {
   ROS_INFO("Updated segmented object list received.");
   // lock for the vector