+++ rail_recognition/include/rail_recognition/ModelGenerator.h
@@ -1,172 +1,122 @@
-//ROS
-#include <ros/ros.h>
+/*!
+ * \file ModelGenerator.h
+ * \brief The grasp model generator node object.
+ *
+ * The grasp model generator allows for generating graspdb models based on registration criteria. An action server is
+ * used to provide the model/grasp demonstration IDs to use during registration.
+ *
+ * \author David Kent, WPI - rctoris@wpi.edu
+ * \author Russell Toris, WPI - rctoris@wpi.edu
+ * \date April 6, 2015
+ */
+
+#ifndef RAIL_PICK_AND_PLACE_MODEL_GENERATOR_H_
+#define RAIL_PICK_AND_PLACE_MODEL_GENERATOR_H_
+
+// RAIL Recognition
+#include "PCLGraspModel.h"
+
+// ROS
 #include <actionlib/server/simple_action_server.h>
-#include <geometry_msgs/PoseArray.h>
 #include <graspdb/graspdb.h>
 #include <rail_pick_and_place_msgs/GenerateModelsAction.h>
-#include <rail_pick_and_place_msgs/GraspWithSuccessRate.h>
-#include <rail_recognition/Model.h>
-#include <sensor_msgs/PointCloud.h>
-#include <sensor_msgs/point_cloud_conversion.h>
-#include <tf/transform_broadcaster.h>
-#include <tf/transform_listener.h>
+#include <ros/ros.h>
 
-//C++
-#include <boost/thread/thread.hpp>
-#include <stdlib.h>
-
-//PCL
-#include <pcl/io/pcd_io.h>
-#include <pcl/kdtree/kdtree_flann.h>
+// PCL
 #include <pcl/point_cloud.h>
 #include <pcl/point_types.h>
-#include <pcl/registration/icp.h>
-#include <pcl_conversions/pcl_conversions.h>
-
-//Point Cloud Filtering Constants
-#define RADIUS .01
-#define NUM_NEIGHBORS 23
-#define DST_THRESHOLD .00075
 
 namespace rail
 {
 namespace pick_and_place
 {
 
+/*!
+ * \class ModelGenerator
+ * \brief The grasp model generator node object.
+ *
+ * The grasp model generator allows for generating graspdb models based on registration criteria. An action server is
+ * used to provide the model/grasp demonstration IDs to use during registration.
+ */
 class ModelGenerator
 {
 public:
-  /**
-   * Constructor
+  /*! If a topic should be created to display debug information such as model point clouds. */
+  static const bool DEFAULT_DEBUG = false;
+
+  /*!
+   * \brief Create a ModelGenerator and associated ROS information.
+   *
+   * Creates a ROS node handle and starts the action server.
    */
   ModelGenerator();
 
-  /**
-  * Register all possible point cloud pairs
-  */
-  void pairwiseRegisterPointclouds();
+  /*!
+   * \brief Cleans up a ModelGenerator.
+   *
+   * Cleans up any connections used by the ModelGenerator.
+   */
+  virtual ~ModelGenerator();
 
-  /**
-  * Registered point clouds using a graph-based approach to successive pairwise registration
-  * @param models models to register with the graph method
-  * @return the number of generated models
-  */
-  int registerPointCloudsGraph(std::vector<Model> models, int maxModelSize, std::vector<int> unusedModelIds);
+  /*!
+   * \brief A check for a valid ModelGenerator.
+   *
+   * This function will return true if the appropriate connections were created successfully during initialization.
+   *
+   * \return True if the appropriate connections were created successfully during initialization.
+   */
+  bool okay() const;
 
 private:
-  //ROS publishers, subscribers, and action servers
-  ros::NodeHandle n;
-
-  actionlib::SimpleActionServer<rail_pick_and_place_msgs::GenerateModelsAction> asGenerateModels;
-
-  graspdb::Client *graspdb;
-
-  //Point clouds
-  sensor_msgs::PointCloud2 baseCloud;
-  sensor_msgs::PointCloud2 targetCloud;
-
-  //tf
-  tf::TransformListener tfListener;
-  tf::TransformBroadcaster tfBroadcaster;
-
-  std::string outputDirectory;
-
-  std::vector<Model> individualGraspModels;
-  std::vector<Model> mergedModels;
-
-  void executeGenerateModels(const rail_pick_and_place_msgs::GenerateModelsGoalConstPtr &goal);
-
-  /**
-   * Run pcl's icp on a base and target point cloud
-   * @param baseCloudPtr pointer to the point cloud to which the target will be transformed
-   * @param targetCloudPtr pointer to the point cloud that will be transformed to the base cloud
-   * @param baseGrasps list of grasps associated with the base point cloud
-   * @param targetGrasps list of grasps associated with the target point cloud
-   * @param resultGrasps pointer to the list of grasps that will be filled with transformed grasps
-   * @return a pointer to the merged point cloud
+  /*!
+   * \brief Main model generation callback.
+   *
+   * The main callback for the model generation action server.
+   *
+   * \param goal The goal specifying the parameters for the model generation.
    */
-  pcl::PointCloud<pcl::PointXYZRGB>::Ptr icpRegistration(pcl::PointCloud<pcl::PointXYZRGB>::Ptr baseCloudPtr,
-    pcl::PointCloud<pcl::PointXYZRGB>::Ptr targetCloudPtr,
-    std::vector<rail_pick_and_place_msgs::GraspWithSuccessRate> baseGrasps,
-    std::vector<rail_pick_and_place_msgs::GraspWithSuccessRate> targetGrasps,
-    std::vector<rail_pick_and_place_msgs::GraspWithSuccessRate> *resultGrasps);
+  void generateModelsCallback(const rail_pick_and_place_msgs::GenerateModelsGoalConstPtr &goal);
 
-  /**
-   * Determine if a point cloud registration is successful based on various score metrics
-   * @param baseCloudPtr pointer to the point cloud to which the target will be transformed
-   * @param targetCloudPtr pointer to the point cloud that will be transformed to the base cloud
+  /*!
+   * \brief Model generation function.
+   *
+   * Attempt to search for valid registration pairs for the given models up to the max model size. Valid models are
+   * saved to the database and a list of IDs is stored in the given vector.
+   *
+   * \param grasp_models The array of grasp models to attempt to generate models for.
+   * \param max_model_size The maximum number of grasps allowed per model.
+   * \param new_model_ids The vector to fill with the new grasp model IDs.
    */
-  bool checkRegistration(pcl::PointCloud<pcl::PointXYZRGB>::Ptr baseCloudPtr,
-    pcl::PointCloud<pcl::PointXYZRGB>::Ptr targetCloudPtr);
+  void generateAndStoreModels(std::vector<PCLGraspModel> &grasp_models, const int max_model_size,
+      std::vector<uint32_t> &new_model_ids);
 
-  /**
-   * Calculate a metric for how successful the registration was based on distance error
-   * @param baseCloudPtr pointer to the point cloud to which the target will be transformed
-   * @param targetCloudPtr pointer to the point cloud that will be transformed to the base cloud
-   * @return a score representing the success of the registration
+  /*!
+   * \brief Check the point cloud registration for the two models.
+   *
+   * Attempt to merge the two models into a single model. If the model passes the criteria, the result model is
+   * filled with the corresponding model.
+   *
+   * \param base The base model to compare to.
+   * \param target The target model to compare against the base model.
+   * \param new_model_ids The resuliting model (if valid)
+   * \return True if the registration meets the criteria.
    */
-  float calculateRegistrationMetricDstError(pcl::PointCloud<pcl::PointXYZRGB>::Ptr baseCloudPtr,
-    pcl::PointCloud<pcl::PointXYZRGB>::Ptr targetCloudPtr);
+  bool registrationCheck(const PCLGraspModel &base, const PCLGraspModel &target, PCLGraspModel &result) const;
 
-  /**
-   * Calculate a metric for how successful the registration was based on overlap
-   * @param baseCloudPtr pointer to the point cloud to which the target will be transformed
-   * @param targetCloudPtr pointer to the point cloud that will be transformed to the base cloud
-   * @return a score representing the success of the registration
-   */
-  float calculateRegistrationMetricOverlap(pcl::PointCloud<pcl::PointXYZRGB>::Ptr baseCloudPtr,
-    pcl::PointCloud<pcl::PointXYZRGB>::Ptr targetCloudPtr, float dstThreshold);
+  /*! The debug flag. */
+  bool debug_, okay_;
+  /*! The grasp database connection. */
+  graspdb::Client *graspdb_;
 
-  /**
-   * Calculate a metric for how successful the registration was based on range of colors in each point cloud
-   * @param baseCloudPtr pointer to the point cloud to which the target will be transformed
-   * @param targetCloudPtr pointer to the point cloud that will be transformed to the base cloud
-   * @return a score representing the success of the registration
-   */
-  float calculateRegistrationMetricColorRange(pcl::PointCloud<pcl::PointXYZRGB>::Ptr baseCloudPtr,
-    pcl::PointCloud<pcl::PointXYZRGB>::Ptr targetCloudPtr);
-
-  /**
-   * Calculate a metric for how successful the registration was based on the respective spatial lengths of each point cloud
-   * @param baseCloudPtr pointer to the point cloud to which the target will be transformed
-   * @param targetCloudPtr pointer to the point cloud that will be transformed to the base cloud
-   * @return a score representing the success of the registration
-   */
-  float calculateRegistrationMetricDistance(pcl::PointCloud<pcl::PointXYZRGB>::Ptr baseCloudPtr,
-    pcl::PointCloud<pcl::PointXYZRGB>::Ptr targetCloudPtr);
-
-  /**
-   * Filters point cloud outliers if they have less neighbors than the neighbor threshold within a given radius
-   * @param cloudPtr pointer to the point cloud to be filtered
-   * @param radius the radius to search within for neighbors
-   * @param numNeighborThreshold minimum number of neighbors required to keep the point
-   */
-  void filterCloudOutliers(pcl::PointCloud<pcl::PointXYZRGB>::Ptr cloudPtr, double radius, int numNeighborThreshold);
-
-  /**
-   * Removes extra points that are within a given threshold of other points to keep the point cloud size manageable
-   * @param cloudPtr pointer to the point cloud to be filtered
-   * @param dstThreshold the minimum distance between neighboring points at which to keep a point
-   */
-  void filterRedundentPoints(pcl::PointCloud<pcl::PointXYZRGB>::Ptr cloudPtr, double dstThreshold);
-
-  /**
-   * Translates a point cloud so that it's average point lies on the origin
-   * @param cloudPtr pointer to the point cloud to be translated
-   * @param grasps pointer to the grasps corresponding to the point cloud to be translated
-   */
-  void translateToOrigin(pcl::PointCloud<pcl::PointXYZRGB>::Ptr cloudPtr,
-    std::vector<rail_pick_and_place_msgs::GraspWithSuccessRate> *grasps);
-
-  /**
-   * Classify a point cloud merge as successful or not, based on a decision tree learned previously
-   * @param overlap metric representing the overlap of the two point clouds
-   * @param maxDstDiff metric representing the difference in maximum spatial lengths of the two point clouds
-   * @param dstError metric representing the error in position of points in the two point clouds
-   * @param avgColorDiff metric representing the difference in the colors of the point clouds
-   */
-  bool classifyMerge(float overlap, float maxDstDiff, float dstError, float avgColorDiff);
+  /*! The public and private ROS node handles. */
+  ros::NodeHandle node_, private_node_;
+  /*! The main model generation action server. */
+  actionlib::SimpleActionServer<rail_pick_and_place_msgs::GenerateModelsAction> as_;
+  /*! The debug topic publisher. */
+  ros::Publisher debug_pc_pub_, debug_poses_pub_;
 };
 
-} // end namespace pick_and_place
-} // end namespace rail
+}
+}
+
+#endif