+++ rail_pick_and_place_tools/CHANGELOG.rst
@@ -2,6 +2,54 @@
 Changelog for package rail_pick_and_place_tools
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.1.3 (2015-04-14)
+------------------
+
+Forthcoming
+-----------
+* changelog updated
+* Contributors: Russell Toris
+
+1.1.2 (2015-04-10)
+------------------
+* docs added for rail_pick_and_place_tools
+* splash screen added
+* new metric trainier
+* graph search used
+* Contributors: Russell Toris
+
+1.1.1 (2015-04-06)
+------------------
+* refactor of RViz GUIs
+* cleanup of grasp collection panel
+* Contributors: Russell Toris
+
+1.1.0 (2015-04-03)
+------------------
+* new launch files
+* grasp model retriever added
+* travis fix
+* travis fix
+* Merge branch 'develop' of github.com:WPI-RAIL/rail_pick_and_place into develop
+* major refactor of model generation
+* Buttons should now correctly handle the case of an unconnected action client - server pair
+* Added confirmation dialog for model removal
+* Contributors: David Kent, Russell Toris
+
+1.0.4 (2015-03-31)
+------------------
+* removed old recognize actions
+* Contributors: Russell Toris
+
+1.0.3 (2015-03-31)
+------------------
+* starting to cleanup model generator
+* refactor of point cloud recognizer
+* Contributors: Russell Toris
+
+1.0.2 (2015-03-30)
+------------------
+
 1.0.1 (2015-03-27)
 ------------------
 * Label change